+++ example/account_cmd_parser_example.dart
@@ -1,6 +1,7 @@
 import 'package:account_cmd_parser/account_cmd_parser.dart';
 
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
+void main(List<String> arguments) {
+  final accountCmdParser = AccountArgParser();
+  Account account = accountCmdParser.parse(arguments: arguments);
+  print(account);
 }