+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.0
-
-- Initial version.
+* Parse command-line arguments to `Account(email, password)` model
+* Hint and parse console inputs to `Account(email, password)` model if there are no runtime command-line arguments passed in